requirements.txt
asgiref==3.2.3
autopep8==1.5.1
boto3==1.12.26
botocore==1.15.26
Django==3.0.1
django-cors-headers==3.2.0
django-countries==6.1.2
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
gunicorn==20.0.4
image==1.5.27
jmespath==0.9.5
Pillow==7.0.0
pycodestyle==2.5.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
